refactoring register
