+++ src/main/java/Kelompok2_RPL/AplikasiKlinik/Pasien.java
@@ -2,15 +2,15 @@ package Kelompok2_RPL.AplikasiKlinik;
 
 import java.time.LocalDate;
 
-import org.springframework.format.annotation.DateTimeFormat;
-
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @AllArgsConstructor
 @Data
+@NoArgsConstructor
 public class Pasien {
     @NotNull
     private String nama;
@@ -23,10 +23,9 @@ public class Pasien {
     @Size(min=4, max=60)
     private String rePassword;
     @NotNull
-    @DateTimeFormat(pattern = "yyyy-MM-dd") // Match the input type="date" format
     private LocalDate tanggalLahir;
     @NotNull
     private String nomorTlp;
-    
+    private String alamat;
     private String jenis;
 }
