--- src/main/java/Kelompok2_RPL/AplikasiKlinik/RegisterRepository.java
@@ -1,8 +1,0 @@
-package Kelompok2_RPL.AplikasiKlinik;
-
-import java.util.Optional;
-
-public interface RegisterRepository {
-    void save(Register pasien) throws Exception;
-    Optional<Register> findByEmail(String email);
-}
