+++ src/main/java/Kelompok2_RPL/AplikasiKlinik/register/RegisterService.java
@@ -1,10 +1,12 @@
-package Kelompok2_RPL.AplikasiKlinik;
+package Kelompok2_RPL.AplikasiKlinik.register;
 
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+
+import Kelompok2_RPL.AplikasiKlinik.Pasien;
 
 @Service
 public class RegisterService {
@@ -14,10 +16,10 @@ public class RegisterService {
     @Autowired
     private PasswordEncoder passwordEncoder;
 
-    public boolean register(Register register ){
+    public boolean register(Pasien register ){
         try{
 
-            Optional<Register> existingPasien =registerRepository.findByEmail(register.getEmail());
+            Optional<Pasien> existingPasien =registerRepository.findByEmail(register.getEmail());
             if(existingPasien.isPresent()){
                 System.out.println("Email already exists: " + register.getEmail());
                 return false;
